Fix type error by reusing component
